Unify the logic to get pod parent info.
