+++ pkg/action/util/util.go
@@ -1,7 +1,6 @@
 package util
 
 import (
-	"encoding/json"
 	"fmt"
 
 	"github.com/turbonomic/kubeturbo/pkg/util"
@@ -100,88 +99,6 @@ func GetNodeFromProperties(kubeClient *client.Clientset, properties []*proto.Ent
 // Given namespace and name, return an identifier in the format, namespace/name
 func BuildIdentifier(namespace, name string) string {
 	return namespace + "/" + name
-}
-
-func GetPod(kubeClient *client.Clientset, namespace, name string) (*api.Pod, error) {
-	return kubeClient.CoreV1().Pods(namespace).Get(name, metav1.GetOptions{})
-}
-
-func parseOwnerReferences(owners []metav1.OwnerReference) (string, string) {
-	for i := range owners {
-		owner := &owners[i]
-		if *(owner.Controller) && len(owner.Kind) > 0 && len(owner.Name) > 0 {
-			return owner.Kind, owner.Name
-		}
-	}
-
-	return "", ""
-}
-
-func GetPodParentInfo(pod *api.Pod) (string, string, error) {
-	//1. check ownerReferences:
-
-	if pod.OwnerReferences != nil && len(pod.OwnerReferences) > 0 {
-		kind, name := parseOwnerReferences(pod.OwnerReferences)
-		if len(kind) > 0 && len(name) > 0 {
-			return kind, name, nil
-		}
-	}
-
-	glog.V(4).Infof("no parent-info for pod-%v/%v in OwnerReferences.", pod.Namespace, pod.Name)
-
-	//2. check annotations:
-	if pod.Annotations != nil && len(pod.Annotations) > 0 {
-		key := "kubernetes.io/created-by"
-		if value, ok := pod.Annotations[key]; ok {
-
-			var ref api.SerializedReference
-
-			if err := json.Unmarshal([]byte(value), &ref); err != nil {
-				err = fmt.Errorf("failed to decode parent annoation:%v", err)
-				glog.Errorf("%v\n%v", err, value)
-				return "", "", err
-			}
-
-			return ref.Reference.Kind, ref.Reference.Name, nil
-		}
-	}
-
-	glog.V(4).Infof("no parent-info for pod-%v/%v in Annotations.", pod.Namespace, pod.Name)
-
-	return "", "", nil
-}
-
-// get grandParent(parent's parent) information of a pod: kind, name
-// If parent does not have parent, then return parent info.
-// Note: if parent kind is "ReplicaSet", then its parent's parent can be a "Deployment"
-func GetPodGrandInfo(kclient *client.Clientset, pod *api.Pod) (string, string, error) {
-	//1. get Parent info: kind and name;
-	kind, name, err := GetPodParentInfo(pod)
-	if err != nil {
-		return "", "", err
-	}
-
-	//2. if parent is "ReplicaSet", check parent's parent
-	if strings.EqualFold(kind, "ReplicaSet") {
-		//2.1 get parent object
-		rs, err := kclient.ExtensionsV1beta1().ReplicaSets(pod.Namespace).Get(name, metav1.GetOptions{})
-		if err != nil {
-			err = fmt.Errorf("Failed to get ReplicaSet[%v/%v]: %v", pod.Namespace, name, err)
-			glog.Error(err.Error())
-			return "", "", err
-		}
-
-		//2.2 get parent's parent info by parsing ownerReferences:
-		// TODO: The ownerReferences of ReplicaSet is supported only in 1.6.0 and afetr
-		if rs.OwnerReferences != nil && len(rs.OwnerReferences) > 0 {
-			gkind, gname := parseOwnerReferences(rs.OwnerReferences)
-			if len(gkind) > 0 && len(gname) > 0 {
-				return gkind, gname, nil
-			}
-		}
-	}
-
-	return kind, name, nil
 }
 
 // check whether parentKind is supported for MovePod/ResizeContainer actions
